move card component and styles into seperate folder
