+++ src/components/FeaturedListing/index.jsx
@@ -8,7 +8,7 @@ import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import Slider from 'react-slick';
 
-import Card from './Card';
+import Card from './Card/Card';
 import data from './data';
 import SectionContainer, {
   StyledSlide,
